api/serializers.py
from django.db import models
from rest_framework import fields, serializers
from .models import Employee

class EmployeeSerializer(serializers.ModelSerializer):
    class Meta:
        model = Employee
        fields = '__all__'