requirements.txt
asgiref==3.4.1
cycler==0.11.0
Django==4.0
django-cors-headers==3.13.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
fonttools==4.33.3
gunicorn==20.1.0
kiwisolver==1.4.2
matplotlib==3.5.2
mysql-connector==2.2.9
mysql-connector-python==8.0.29
numpy==1.22.3
openpyxl==3.0.9
packaging==21.3
pandas==1.4.2
Pillow==9.0.0
pyparsing==3.0.9
PyPDF2==1.27.12
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tkinter-math==0.2.4
tzdata==2021.5
whitenoise==6.2.0
